+++ bundler/analyzer/lib/transpileImportTokensToCJS/test/index.js
@@ -638,6 +638,80 @@ runTest(
     { type: "punctuator", value: "}" },
     { type: "punctuator", value: ")" },
     { type: "punctuator", value: ";" },
+  ]
+);
+
+runTest(
+  "Import - dynamic inside arrow function call",
+  transpileImportTokensToCJSPreprocessor([
+    { type: "identifier", value: "foo" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"/\"" },
+    { type: "punctuator", value: "," },
+    { type: "whitespace", value: " " },
+    { type: "identifier", value: "ctx" },
+    { type: "whitespace", value: " " },
+    { type: "punctuator", value: "=>" },
+    { type: "whitespace", value: " " },
+    { type: "keyword", value: "import" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"./assets/js/baz.js\"" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ";" }
+  ]),
+  [
+    { type: "identifier", value: "foo" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"/\"" },
+    { type: "punctuator", value: "," },
+    { type: "identifier", value: "ctx" },
+    { type: "punctuator", value: "=>" },
+    { type: "identifier", value: "requireByHttp" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"./assets/js/baz.js\"" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ";" }
+  ]
+);
+
+runTest(
+  "Import - dynamic inside arrow function call (parenthesized param)",
+  transpileImportTokensToCJSPreprocessor([
+    { type: "identifier", value: "foo" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"/\"" },
+    { type: "punctuator", value: "," },
+    { type: "whitespace", value: " " },
+    { type: "punctuator", value: "(" },
+    { type: "identifier", value: "ctx" },
+    { type: "punctuator", value: ")" },
+    { type: "whitespace", value: " " },
+    { type: "punctuator", value: "=>" },
+    { type: "whitespace", value: " " },
+    { type: "keyword", value: "import" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"./assets/js/bar.js\"" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ";" }
+  ]),
+  [
+    { type: "identifier", value: "foo" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"/\"" },
+    { type: "punctuator", value: "," },
+    { type: "punctuator", value: "(" },
+    { type: "identifier", value: "ctx" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: "=>" },
+    { type: "identifier", value: "requireByHttp" },
+    { type: "punctuator", value: "(" },
+    { type: "string", value: "\"./assets/js/bar.js\"" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ")" },
+    { type: "punctuator", value: ";" }
   ],
   true
 );